lib/operations: Range is int, not int64

+++ cmd/index.go
@@ -1,7 +1,7 @@
 package cmd
 
 // Range is a text serialisable width.
-type Range struct{ Min, Max int64 }
+type Range struct{ Min, Max int }
 
 // Operation is a text serialisable lib.Operation.
 type Operation struct {
